break import cycle loop with a concrete function field
