+++ internal/consul/kv_test.go
@@ -5,19 +5,18 @@ import (
 
 	consulApi "github.com/hashicorp/consul/api"
 	"github.com/stretchr/testify/assert"
-	"github.com/trento-project/trento/internal/consul/mocks"
 )
 
 func TestKVMaps(t *testing.T) {
 
-  kvpairs := consulApi.KVPairs{
-		&consulApi.KVPair{Key: "/trento/", Value: []byte("")},
-		&consulApi.KVPair{Key: "/trento/env1/", Value: []byte("")},
+	kvPairs := consulApi.KVPairs{
+		&consulApi.KVPair{Key: "/trento/", Value: []byte("foo")},
+		&consulApi.KVPair{Key: "/trento/env1/", Value: []byte("bar")},
 		&consulApi.KVPair{Key: "/trento/env1/landscapes/", Value: []byte("")},
 		&consulApi.KVPair{Key: "/trento/env1/landscapes/land1/", Value: []byte("")},
 		&consulApi.KVPair{Key: "/trento/env1/landscapes/land1/sapsystems/", Value: []byte("")},
 		&consulApi.KVPair{Key: "/trento/env1/landscapes/land1/sapsystems/sys1/", Value: []byte("")},
-    &consulApi.KVPair{Key: "/trento/env1/landscapes/land1/sapsystems/sys1/name", Value: []byte("name")},
+		&consulApi.KVPair{Key: "/trento/env1/landscapes/land1/sapsystems/sys1/name", Value: []byte("name")},
 		&consulApi.KVPair{Key: "/trento/env1/landscapes/land2/", Value: []byte("")},
 		&consulApi.KVPair{Key: "/trento/env1/landscapes/land2/sapsystems/", Value: []byte("")},
 		&consulApi.KVPair{Key: "/trento/env1/landscapes/land2/sapsystems/sys2/", Value: []byte("")},
@@ -28,15 +27,12 @@ func TestKVMaps(t *testing.T) {
 		&consulApi.KVPair{Key: "/trento/env2/landscapes/land3/sapsystems/sys3/", Value: []byte("")},
 	}
 
-	kv := new(KV)
+	kv := &kv{list: func(prefix string, q *consulApi.QueryOptions) (consulApi.KVPairs, *consulApi.QueryMeta, error) {
+		return kvPairs, nil, nil
+	}}
+	result, _ := kv.Maps("", "")
 
-	consulApi.On("KV").Return(kv)
+	expectedMap := map[string]interface{}{"/trento/": "foo"}
 
-	kv.On("List", "/trento", (*consulApi.QueryOptions)(nil)).Return(kvpairs, nil, nil)
-
-  result, _ := Maps(kv, "/trento/", "/trento")
-
-  expectedMap := map[string]interace{"a": 1}
-
-  assert.Equal(t, result, expectedMap)
+	assert.Equal(t, expectedMap, result)
 }
